feat: alert SweetAlert2 적용 refactor : context API로 리팩터링
