+++ src/context/PokemonProvider.jsx
@@ -1,0 +1,40 @@
+import React, { createContext, useContext, useState } from "react";
+import Swal from "sweetalert2";
+import "sweetalert2/dist/sweetalert2.min.css";
+
+const PokemonContext = createContext();
+
+export function usePokemon() {
+  return useContext(PokemonContext);
+}
+
+export function PokemonProvider({ children }) {
+  const [selectedPokemon, setSelectedPokemon] = useState([]);
+
+  const addPokemon = (pokemon) => {
+    if (selectedPokemon.length >= 6) {
+      //   alert("최대 6개의 포켓몬만 선택할 수 있습니다.");
+      Swal.fire({
+        html: '<div style="font-size: 25px;">최대 6개의 포켓몬만 선택할 수 있습니다.</div>',
+      });
+      return;
+    }
+    if (!selectedPokemon.some((p) => p.id === pokemon.id)) {
+      setSelectedPokemon((prevSelected) => [...prevSelected, pokemon]);
+    }
+  };
+
+  const removePokemon = (pokemon) => {
+    setSelectedPokemon((prevSelected) =>
+      prevSelected.filter((p) => p.id !== pokemon.id)
+    );
+  };
+
+  return (
+    <PokemonContext.Provider
+      value={{ selectedPokemon, addPokemon, removePokemon }}
+    >
+      {children}
+    </PokemonContext.Provider>
+  );
+}
